backend/app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aws-lambda-powertools>=3.15.1",
    "boto3>=1.39.0",
    "fastapi>=0.115.14",
    "mangum>=0.19.0",
    "pillow>=11.3.0",
    "pip>=25.1.1",
    "psycopg2-binary>=2.9.10",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
    "python-dateutil>=2.8.2",
    "cryptography>=45.0.5",
    "pydantic>=1.10.22",
    "typing-extensions>=4.14.0",
]
